requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
PyJWT
pydantic
resend
pytest
pytest-asyncio
httpx
